SPFA-137: Modify router to align on current code
Modify the router. Created a method in controller and update app.js to align with existing code
